src/requirements.txt
async-timeout==4.0.3
asyncpg==0.29.0
fastapi==0.110.1
importlib-metadata==7.0.1
llvmlite==0.42.0
minio==7.2.5
Pillow==10.3.0
pydantic==1.10.12
python-dotenv==1.0.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
python_bcrypt==0.3.2
python_jose==3.3.0
python-multipart==0.0.9
pytz==2023.3.post1
requests==2.31.0
starlette==0.37.2
SQLAlchemy==1.4.52
urllib3==2.0.7